// rick-morty/src/features/common/services/rick-morty/types.ts
export type TCharacter = {
  id: string;
  name: string;
  status: string;
  species: string;
  type: string;
  gender: string;
  origin: {
    name: string;
    url: string;
  };
  location: {
    name: string;
    url: string;
  };
  image: string;
  episode: string[];
  url: string;
  created: string;
};

export type TResponse = {
  info: {
    count: number;
    pages: number;
    next: string | null;
    prev: string | null;
  };

  results: TCharacter[];
};